AB#108987: Add publisher and datePublished to the root data entity (#224)

+++ app/HutchAgent/HostedServices/JobPollingHostedService.cs
@@ -65,6 +65,7 @@ public class JobPollingHostedService : BackgroundService
 
   private async Task UploadResults()
   {
+    if (_wfexsJobService is null) throw new NullReferenceException("_wfexsJobService instance not available");
     var finishedJobs = await _wfexsJobService.ListFinishedJobs();
     foreach (var job in finishedJobs)
     {
@@ -83,6 +84,7 @@ public class JobPollingHostedService : BackgroundService
           job.WfexsRunId)
       );
 
+      if (_resultsStoreWriter is null) throw new NullReferenceException("_resultsStoreWriter instance not available");
       if (await _resultsStoreWriter.ResultExists(pathToUploadInfo.Name))
       {
         _logger.LogInformation($"{pathToUploadInfo.Name} already exists in S3.");
@@ -108,6 +110,7 @@ public class JobPollingHostedService : BackgroundService
 
   private async Task MergeResults()
   {
+    if (_wfexsJobService is null) throw new NullReferenceException("_wfexsJobService instance not available");
     var finishedJobs = await _wfexsJobService.ListFinishedJobs();
     foreach (var job in finishedJobs)
     {
@@ -116,7 +119,7 @@ public class JobPollingHostedService : BackgroundService
         "wfexs-backend-test_WorkDir",
         job.WfexsRunId);
       var sourceZip = Path.Combine(jobWorkDir, "outputs", $"{job.WfexsRunId}.zip");
-      var pathToMetadata = Path.Combine(job.UnpackedPath, "ro-crate-metadata.json");
+      var pathToMetadata = Path.Combine(job.UnpackedPath); // directory path to ro-crate-metadata.json
       var mergeDirInfo = new DirectoryInfo(job.UnpackedPath);
       var mergeDirParent = mergeDirInfo.Parent;
       var mergedZip = Path.Combine(mergeDirParent!.FullName, $"{mergeDirInfo.Name}-merged.zip");
@@ -128,11 +131,13 @@ public class JobPollingHostedService : BackgroundService
         continue;
       }
 
+      if (_crateMergerService is null) throw new NullReferenceException("_crateMergerService instance not available");
       _crateMergerService.MergeCrates(sourceZip, job.UnpackedPath);
       _crateMergerService.DeleteContainerImages(pathToContainerImagesDir);
       _crateMergerService.UpdateMetadata(pathToMetadata);
       _crateMergerService.ZipCrate(job.UnpackedPath);
 
+      if (_resultsStoreWriter is null) throw new NullReferenceException("_resultsStoreWriter instance not available");
       if (await _resultsStoreWriter.ResultExists(mergedZip))
       {
         _logger.LogInformation($"Merged Crate {mergedZip} already exists in results store.");
@@ -149,6 +154,7 @@ public class JobPollingHostedService : BackgroundService
   /// </summary>
   private async Task CheckJobsFinished()
   {
+    if (_wfexsJobService is null) throw new NullReferenceException("_wfexsJobService instance not available");
     var unfinishedJobs = await _wfexsJobService.List();
     unfinishedJobs = unfinishedJobs.FindAll(x => !x.RunFinished);
